Move map utilities to helpers folder
